Group coupled maximizer value types
Keep package-private acquisition and resource schemas beside the values that consume them.
