--- src/net/sourceforge/kolmafia/maximizer/AcquisitionMethod.java
@@ -1,12 +1,0 @@
-package net.sourceforge.kolmafia.maximizer;
-
-enum AcquisitionMethod {
-  ACCESSIBLE,
-  CREATE,
-  NPC_BUY,
-  FOLD,
-  PULL,
-  PULL_FOLD,
-  STORAGE_BUY,
-  MALL_BUY
-}
