+++ src/net/sourceforge/kolmafia/maximizer/ResourceUsage.java
@@ -2,6 +2,7 @@ package net.sourceforge.kolmafia.maximizer;
 
 import java.util.List;
 import java.util.Map;
+import java.util.function.ToIntFunction;
 import net.sourceforge.kolmafia.AdventureResult;
 import net.sourceforge.kolmafia.equipment.Slot;
 
@@ -96,3 +97,9 @@ final class ResourceUsage implements Comparable<ResourceUsage> {
     return Integer.compare(other.costs.length, this.costs.length);
   }
 }
+
+record SharedResource(
+    String name,
+    int limit,
+    ToIntFunction<Map<Slot, AdventureResult>> equipmentCost,
+    ToIntFunction<String> itemCost) {}
